+++ kartoteka-ui-gtk/src/main.rs
@@ -5,6 +5,7 @@ mod config;
 mod github;
 mod secret_store;
 mod ui;
+mod webdav;
 
 use gtk4::prelude::*;
 use libadwaita as adw;